Use CIELAB 1994 metric instead of Eucledean distance metric
There's also a preliminary implementation of the CIELAB 2000 metric, but
it looks like it gives some wrong results for some colors, so it probably
has bugs. This commit uses the 1994 version.

Also, cie94 really loves the green component. This is why the reference value
for the yellow color is changed from (1,1,0) to (1,200/255,0). Empirically
better results.