Make a "list of strings" for the cmd the default behavior
This is list of strings is passed to LinuxPty and handled from there.
If "cmd" is a string, it is split with shlex.split to account for single
and double quotes. These changes are also reflected in the exec command,
where a `"shell_cmd"` argument is split with shlex, and a `"cmd"`
argument is used as-is.